split trial context in two
